# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de base en laravel",
        "description": "MDN",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://mdn-rdv.com/api"
        }
    ],
    "paths": {
...
        "/admin/recepts/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__RECEPTIONISTE"
                ],
                "summary": "Enregistre un réceptionniste",
                "description": "",
                "operationId": "c97eeed36daa420f50d7bcbfd41ef477",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mulamba"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "0852277379"
                                    },
                                    "role": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/recepts/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__RECEPTIONISTE"
                ],
                "summary": "Renvoie les informations d'un réceptionniste par son ID",
                "operationId": "b606fea83dcb0cbf50ba9bacc95319b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/recepts/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__RECEPTIONISTE"
                ],
                "summary": "Editer les informations d'un réceptionniste",
                "operationId": "b176c005c5fc77febb5d00b3d4f9c209",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Mulamba"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "0852277379"
                                    },
                                    "role": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/recepts/activate/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__RECEPTIONISTE"
                ],
                "summary": "Active ou désactive l'état d'un réceptionniste par son ID",
                "operationId": "f4527aaf32777ed5cbbd62463209f78a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/recepts/delete": {
            "post": {
                "tags": [
                    "ADMINISTRATION__RECEPTIONISTE"
                ],
                "summary": "Supprimer un receptionniste",
                "operationId": "2f95df09600a32611b16ec24051bf4f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            },
            "0": []
        }
    }
}